analysis/datastub/SymbolInfo: Get debug file using gdb
For libc and other installed libaries a separate package can be installed to access
debug symbols. But for this a separate elffile is stored with an arbitrary name.
gdb does automagically detect this and loads the correct files.
If nm fails to load any symbols, gdb is used as a fallback to find the debug
elffile.