pt/addrtrace: Exit if no log_addr, except phase 1
For any accessed virtual address, DATA tries to resolve the matching
logical address. There seems to be a general memory range that is not
mapped to any image, but used generally.

DATA allows such an event to occur in phase 1, but will fail if it
occurs in any other phase. Experiments have shown that these accessed
addresses do not differ between different runs.